current state of experimenting with tsondb integration
Currently, the ModelContainer is fully local and the data root is set to /data, which is only for testing purposes. The data root should be defined in optolith-data and then ModelContainer somehow needs to be split up to support generating outputs without providing a data root.

Naming inspiration was drawn from SwiftData.